refactor by cr

+++ server/coordinator/procedure/ddl/droppartitiontable/drop_partition_table.go
@@ -288,14 +288,18 @@ func dropDataTablesCallback(event *fsm.Event) {
 			continue
 		}
 
-		shardVersionUpdate, err := ddl.GetTableShardVersion(table, params.ClusterMetadata, shardVersions)
+		shardVersionUpdate, shardExists, err := ddl.BuildShardVersionUpdate(table, params.ClusterMetadata, shardVersions)
 		if err != nil {
-			_, err := params.ClusterMetadata.DropTableMetadata(req.ctx, req.schemaName(), tableName)
-			if err != nil {
-				procedure.CancelEventWithLog(event, err, "drop table metadata", zap.String("tableName", tableName))
-				return
-			}
 			log.Warn("get shard version by table", zap.String("tableName", tableName), zap.Error(err))
+			// If the shard corresponding to this table does not exist, it means that the actual table creation failed.
+			// In order to ensure that the table can be deleted normally, we need to directly delete the metadata of the table.
+			if !shardExists {
+				_, err := params.ClusterMetadata.DropTableMetadata(req.ctx, req.schemaName(), tableName)
+				if err != nil {
+					procedure.CancelEventWithLog(event, err, "drop table metadata", zap.String("tableName", tableName))
+					return
+				}
+			}
 			continue
 		}
